protocol-v2-types: group related code
